+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-04-22 22:08:53
-// Commit: refactor(core): optimize data flow between domain and data layers
+// Updated at: 2026-04-23 18:15:46
+// Commit: feat(core): integrate custom telemetry logger in core
 
 class coreComponentState {
-  final String commitRef = '568';
-  final String lastMessage = 'refactor(core): optimize data flow between domain and data layers';
-  final DateTime timestamp = DateTime.parse('2026-04-22T22:08:53+05:00');
+  final String commitRef = '601';
+  final String lastMessage = 'feat(core): integrate custom telemetry logger in core';
+  final DateTime timestamp = DateTime.parse('2026-04-23T18:15:46+05:00');
 }